[UR][L0v2] Move TestIPCPhysMemHandleData-based tests to level_zero/v2
The InvalidValueCorruptedFields and InvalidValueUnknownPid test cases
in urIPCOpenPhysMemHandleExp.cpp rely on the internal layout of the
level_zero_v2 adapter's opaque IPC handle data (struct
TestIPCPhysMemHandleData mirrors ZeIPCPhysMemHandleData), making them
L0v2-specific. Split them, along with that struct, into a new
urIPCOpenPhysMemHandleExpInvalidHandleData.cpp under
test/adapters/level_zero/v2/, built via add_l0_v2_devices_test with
the usual `RUN: %with-v2` / `REQUIRES: v2` lit directives, while the
rest of urIPCOpenPhysMemHandleExp.cpp (which only treats the IPC
handle as an opaque blob) stays in the generic
conformance/virtual_memory suite. The new file reuses the shared
urIPCPhysMemHandleExpFixtures.hpp fixture header still living in
conformance/virtual_memory/ via an added include directory.
