+++ unified-runtime/test/adapters/level_zero/v2/CMakeLists.txt
@@ -66,6 +66,13 @@ add_l0_v2_devices_test(memory_residency
     ${PROJECT_SOURCE_DIR}/source/adapters/level_zero/ur_level_zero.cpp
 )
 
+add_l0_v2_devices_test(urIPCOpenPhysMemHandleExpInvalidHandleData
+    urIPCOpenPhysMemHandleExpInvalidHandleData.cpp
+)
+target_include_directories(urIPCOpenPhysMemHandleExpInvalidHandleData-test PRIVATE
+    ${UR_CONFORMANCE_TEST_DIR}/virtual_memory
+)
+
 add_l0_v2_devices_test(batched_queue
     batched_queue_test.cpp
    ${PROJECT_SOURCE_DIR}/source/adapters/level_zero/adapter.cpp
